Move tabs to own folder
