+++ components/planTrip/PlanTrip.js
@@ -1,5 +1,5 @@
 import { View, Text } from 'react-native';
-import styles from '../styles/App.style'
+import styles from '../../styles/App.style'
 
 function PlanTrip() {
     return (
